move permissionlevel and packageaction to separate class
